requirements.txt
astroid==2.11.2
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
dill==0.3.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauthlib==3.2.0
platformdirs==2.5.1
pycodestyle==2.8.0
pylint==2.13.4
python-decouple==3.6
requests==2.27.1
requests-oauthlib==1.3.1
toml==0.10.2
tomli==2.0.1
tweepy==4.8.0
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
